Office Testing
Office can get stuck in a state of prompting for a login even after volume license. I was unable to resolve it with this script, but it does not appear on a fresh install. Collaborators can still sign into Office but won't have a 365 license. If the volume license is there it will still be activated on their login.